+++ client/src/utils/axiosRequest.js
@@ -21,6 +21,7 @@ export async function postSignIn(data) {
     const res = await client.post('/login', data);
     setAccessToken(res.data.accessToken);
     setRefreshToken(res.data.refreshToken);
+    alert('로그인이 성공했습니다');
   } catch (err) {
     alert(err);
   }
@@ -37,11 +38,10 @@ export async function postQuestion(data) {
   }
 }
 
-const refresh_token = getRefreshToken();
 export async function getAccessWithRefresh() {
   try {
     const res = await authenticClient.get('/refresh', {
-      headers: { Refresh: `${refresh_token}` },
+      headers: { Refresh: `${getRefreshToken()}` },
     });
     setAccessToken(res.headers.authorization);
     return res.headers.authorization;
@@ -56,7 +56,7 @@ export async function getAccessWithRefresh() {
 export async function postLogout() {
   try {
     const res = await authenticClient.post('/auths/logout', {
-      headers: { Authorization: `${refresh_token}` },
+      headers: { Authorization: `${getRefreshToken()}` },
     });
     removeAccessToken();
     removeRefreshToken();
@@ -67,7 +67,7 @@ export async function postLogout() {
 
 export async function getUsersActivity(activity, id, page, size) {
   const result = await authenticClient.get(
-    `/users/${id}/${activity}?page=${page}&size=${size}`
+    `/users/${id}/user-${activity}?page=${page}&size=${size}`
   );
   return result;
 }
